pyproject.toml
[tool.poetry]
name = "go-outside-bot"
version = "0.1.0"
description = ""
authors = ["nwunderly <nwunder.dev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
disnake = "^2.9.1"
python-dotenv = "^1.0.1"
black = "^24.2.0"
isort = "^5.13.2"
loguru = "^0.7.2"
disnake-jishaku = "^2.6.5"
psutil = "^5.9.8"
pygit2 = "^1.14.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
